Move hasBy* methods into a shared trait
